+++ kloch/config.py
@@ -7,6 +7,7 @@ import logging
 import os
 from pathlib import Path
 from typing import Dict
+from typing import List
 from typing import Optional
 from typing import Union
 
@@ -14,10 +15,16 @@ import yaml
 
 LOGGER = logging.getLogger(__name__)
 
-KLOCH_CONFIG_ENV_VAR = "KLOCH_CONFIG_PATH"
+KLOCH_CONFIG_PREFIX = "KLOCH_CONFIG"
+
+KLOCH_CONFIG_ENV_VAR = f"{KLOCH_CONFIG_PREFIX}_PATH"
 """
 Environment variable that must specify a file path to an existing configuration file.
 """
+
+
+def _cast_list(src_str: str) -> List[str]:
+    return src_str.split(",")
 
 
 @dataclasses.dataclass
@@ -26,13 +33,27 @@ class KlochConfig:
     Configure kloch using a simple key/value pair dataclass system.
     """
 
+    launcher_plugins: List[str] = dataclasses.field(
+        default_factory=list,
+        metadata={
+            "documentation": (
+                "A list of importable python module names containing new launchers to support.\n\n"
+                "If specified in environment variable, this must be a comma-separated list of str like ``module1,module2,module3``"
+            ),
+            "environ": f"{KLOCH_CONFIG_PREFIX}_launcher_plugins".upper(),
+            "environ_cast": _cast_list,
+        },
+    )
+
     cli_logging_format: str = dataclasses.field(
         default="{levelname: <7} | {asctime} [{name}] {message}",
         metadata={
             "documentation": (
                 "Formatting to use for all logged messages. See python logging module documentation.\n"
                 "The tokens must use the ``{`` style."
-            )
+            ),
+            "environ": f"{KLOCH_CONFIG_PREFIX}_cli_logging_format".upper(),
+            "environ_cast": str,
         },
     )
 
@@ -43,14 +64,11 @@ class KlochConfig:
                 "Logging level to use if None have been specified.\n"
                 "Can be an int or a level name as string as long as it is understandable"
                 " by ``logging.getLevelName``."
-            )
+            ),
+            "environ": f"{KLOCH_CONFIG_PREFIX}_cli_logging_default_level".upper(),
+            "environ_cast": str,
         },
     )
-
-    def __post_init__(self):
-        self.cli_logging_default_level = logging.getLevelName(
-            self.cli_logging_default_level
-        )
 
     @classmethod
     def from_file(cls, file_path: Path) -> "KlochConfig":
@@ -62,14 +80,34 @@ class KlochConfig:
         return cls(**asdict)
 
     @classmethod
-    def from_environment(cls) -> Optional["KlochConfig"]:
+    def from_environment(cls) -> "KlochConfig":
         """
         Generate an instance from a serialized file specified in an environment variable.
         """
         environ = os.getenv(KLOCH_CONFIG_ENV_VAR)
-        if not environ:
-            return None
-        return cls.from_file(Path(environ))
+
+        asdict = {}
+        if environ:
+            base = cls.from_file(Path(environ))
+            asdict = dataclasses.asdict(base)
+
+        for field in dataclasses.fields(cls):
+            env_var_name = field.metadata["environ"]
+            env_var_value = os.getenv(env_var_name)
+            if env_var_value is not None:
+                value = field.metadata["environ_cast"](env_var_value)
+                asdict[field.name] = value
+
+        return cls(**asdict)
+
+    @classmethod
+    def get_field(cls, field_name: str) -> Optional[dataclasses.Field]:
+        """
+        Return the dataclass field that match the given name else None.
+        """
+        fields = dataclasses.fields(cls)
+        field = [field for field in fields if field.name == field_name]
+        return field[0] if field else None
 
 
 def get_config() -> KlochConfig:
@@ -81,5 +119,4 @@ def get_config() -> KlochConfig:
     Returns:
         a new config instance
     """
-    config = KlochConfig.from_environment()
-    return config or KlochConfig()
+    return KlochConfig.from_environment()