+++ kloch/launchers/__init__.py
@@ -6,24 +6,25 @@ from .base import BaseLauncher
 from .base import BaseLauncherSerialized
 from .base import BaseLauncherFields
 
-from .rezenv import RezEnvLauncher
-from .rezenv import RezEnvLauncherSerialized
+
 from .system import SystemLauncher
 from .system import SystemLauncherSerialized
 from .python import PythonLauncher
 from .python import PythonLauncherSerialized
 
+from ._plugins import check_launcher_plugins
+
 from ._get import get_available_launchers_classes
 from ._get import get_launcher_class
 from ._get import get_available_launchers_serialized_classes
 from ._get import get_launcher_serialized_class
+from ._get import is_launcher_plugin
 
 from ._serialized import LauncherSerializedDict
 from ._serialized import LauncherSerializedList
 
-LAUNCHERS = [
+_BUILTINS_LAUNCHERS = [
     BaseLauncher,
-    RezEnvLauncher,
     SystemLauncher,
     PythonLauncher,
 ]
@@ -31,9 +32,8 @@ LAUNCHERS = [
 List of launchers class implementation, including the base one.
 """
 
-LAUNCHERS_SERIALIZED = [
+_BUILTINS_LAUNCHERS_SERIALIZED = [
     BaseLauncherSerialized,
-    RezEnvLauncherSerialized,
     SystemLauncherSerialized,
     PythonLauncherSerialized,
 ]