+++ kloch/launchers/base/_serialized.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from typing import Dict
 from typing import List
 from typing import Type
+from typing import TypeVar
 from typing import Union
 
 from kloch._dictmerge import refacto_dict
@@ -55,16 +56,16 @@ def resolve_environ(environ: Dict[str, Union[str, List[str]]]) -> Dict[str, str]
 
 
 class _MergeableSystemEnviron(MergeableDict):
-    def __init__(self):
-        environ = os.environ.copy()
-        super().__init__({"environ": environ})
-
     @classmethod
     def get_merge_rule(cls, key: str) -> MergeRule:
         # we force to always append the environ keys even if not token
         if cls.resolve_key_tokens(key) == "environ":
             return MergeRule.append
         return super().get_merge_rule(key)
+
+
+def _get_mergeable_system_environ():
+    return _MergeableSystemEnviron({"environ": os.environ.copy()})
 
 
 # we use a dataclass over enum because we need inheritance
@@ -145,6 +146,9 @@ class BaseLauncherFields:
         return list(dataclasses.fields(cls))
 
 
+T = TypeVar("T", bound="BaseLauncherSerialized")
+
+
 class BaseLauncherSerialized(MergeableDict):
     """
     A BaseLauncher instance as a serialized dict object.
@@ -174,6 +178,18 @@ class BaseLauncherSerialized(MergeableDict):
         "An abstract launcher that whose purpose is to be merged with other launchers."
     )
     description = "This launcher is never launched and is simply merged with other launchers defined in the profile."
+
+    def __add__(self: T, other: T) -> T:
+        """
+        Returns:
+            new instance with deepcopied structure.
+        """
+        if not isinstance(other, BaseLauncherSerialized):
+            raise TypeError(
+                f"Cannot concatenate object of type {type(other)} with {type(self)}"
+            )
+        # XXX: we override so the class type is defined by the right member of the + operation
+        return other.__class__(super().__add__(other))
 
     @abc.abstractmethod
     def validate(self):
@@ -205,7 +221,7 @@ class BaseLauncherSerialized(MergeableDict):
         use_system_environ = self.fields.merge_system_environ
 
         if self.get(use_system_environ, True):
-            base = self.__class__(_MergeableSystemEnviron() + self)
+            base = self.__class__(_get_mergeable_system_environ() + self)
             # need to disable else infinite recursion
             base[use_system_environ] = False
             resolved = base.resolved()