+++ tests/data/plugins-rerr/kloch_rerr.py
@@ -1,0 +1,5 @@
+SOME_DUMBASS_CONSTANT = True
+
+
+class ClearlyNotALauncher:
+    useless = True