+++ tests/test_config.py
@@ -11,20 +11,27 @@ def test__KlochConfig():
     kloch.config.KlochConfig()
 
     config = kloch.config.KlochConfig(cli_logging_default_level="DEBUG")
-    assert config.cli_logging_default_level == logging.DEBUG
-    config = kloch.config.KlochConfig(cli_logging_default_level=logging.DEBUG)
     assert config.cli_logging_default_level == "DEBUG"
+    config = kloch.config.KlochConfig(cli_logging_default_level=logging.DEBUG)
+    assert config.cli_logging_default_level == logging.DEBUG
+
+    field = kloch.config.KlochConfig.get_field("cli_logging_default_level")
+    assert field.name == "cli_logging_default_level"
 
 
 def test__KlochConfig__from_environment(monkeypatch, data_dir):
     config = kloch.config.KlochConfig.from_environment()
-    assert config is None
+    assert config == kloch.config.KlochConfig()
 
     config_path = data_dir / "config-blaj.yml"
     monkeypatch.setenv(kloch.config.KLOCH_CONFIG_ENV_VAR, str(config_path))
     config = kloch.config.KlochConfig.from_environment()
-    assert config
-    assert config.cli_logging_default_level == logging.WARNING
+    assert config.cli_logging_default_level == "WARNING"
+    assert config.cli_logging_format == "{levelname: <7}: {message}"
+
+    monkeypatch.setenv("KLOCH_CONFIG_CLI_LOGGING_DEFAULT_LEVEL", "ERROR")
+    config = kloch.config.KlochConfig.from_environment()
+    assert config.cli_logging_default_level == "ERROR"
     assert config.cli_logging_format == "{levelname: <7}: {message}"
 
 
@@ -38,3 +45,5 @@ def test__KlochConfig__from_file(data_dir):
 def test__KlochConfig__documentation():
     for field in dataclasses.fields(kloch.config.KlochConfig):
         assert field.metadata.get("documentation")
+        assert field.metadata.get("environ")
+        assert field.metadata.get("environ_cast")